better handling of SIGTERM in continuous mode
