+++ main.go
@@ -139,31 +139,6 @@ func init() {
 	checksum = []byte{0x00}
 }
 
-func continuous() {
-	log.WithFields(log.Fields{"func": "continuous"}).Infof("Starting continuous mode with Config: %+v", brf)
-
-	startRestAPI()
-
-	success := false
-	for {
-		based = strconv.FormatInt(time.Now().Unix(), 10)
-		switch brf.InfraService {
-		case INFRA_SERVICE_ZK:
-			success = backupZK()
-		case INFRA_SERVICE_ETCD:
-			success = backupETCD()
-		case INFRA_SERVICE_CONSUL:
-			success = backupCONSUL()
-		default:
-			log.WithFields(log.Fields{"func": "continuous"}).Fatal(fmt.Sprintf("Infra service %s unknown or not yet supported", brf.InfraService))
-		}
-		if !success {
-			log.WithFields(log.Fields{"func": "continuous"}).Fatal("Backup was not successfull!")
-		}
-		log.WithFields(log.Fields{"func": "continuous"}).Debugf("Sleeping %d seconds", polltime)
-		time.Sleep(time.Duration(polltime) * time.Second)
-	}
-}
 
 func processop() bool {
 	success := false
